@mmoayyed 2023-05-24
Please see: 

> An implementation of the SAXParserFactory class is NOT guaranteed to be thread safe. It is up to the user application to make sure about the use of the SAXParserFactory from more than one thread. Alternatively the application can have one instance of the SAXParserFactory per thread. An application can use the same instance of the factory to obtain one or more instances of the SAXParser provided the instance of the factory isn't being used in more than one thread at a time.